réorganisation de la partie C++ (nouvelle commande pour tout compiler : g++ -o compute.exe cpp/*.cpp cpp/data/*.cpp cpp/io/*.cpp) et nettoyage des fichiers suivis
