+++ .gitignore
@@ -5,3 +5,6 @@
 /results
 /data/topography/alosalpscustomLCC100m.asc
 /data/passes
+**/*exe
+**/*.pyc
+compute
