+++ albertville.yaml
@@ -4,7 +4,7 @@ input_files:
   airfield_file: ./data/airfields/albertville4326.csv
   topography_file: ./data/topography/alosalpscustomLCC100m.asc
   result_folder: ./results/albertville
-  compute: ./compute
+  compute: ./compute.exe
   # compute: ./compute.exe for windows
   mapcssTemplate: ./templates/mapcss.mapcss
 
@@ -18,7 +18,7 @@ glide_parameters:
   circuit_height: 250
 
 calculation_parameters:
-  max_altitude: 1000
+  max_altitude: 4000
 
 rendering:
   contour_height: 100
