+++ cpp/io/Params.h
@@ -1,0 +1,22 @@
+#ifndef PARAMS_H
+#define PARAMS_H
+
+#include <cstddef>
+#include <string>
+#include <stdexcept>
+#include <iostream>
+#include <algorithm>
+using namespace std;
+
+class Params {
+    public:
+        size_t global_ncols,global_nrows;
+        float homex, homey, finesse, distSol, securite, 
+            nodataltitude,cellsize_m,cellsize_over_finesse, 
+            xllcorner, yllcorner;
+        string output_path, topology, exportPasses;
+
+        Params(int argc, char* argv[]);
+};
+
+#endif // PARAMS_H
